+++ bugsnag-android-core/src/main/java/com/bugsnag/android/BadResponseException.java
@@ -8,7 +8,6 @@ import java.util.Locale;
  * @deprecated use {@link DeliveryFailureException} instead
  */
 @Deprecated
-@ThreadSafe
 public class BadResponseException extends Exception {
     private static final long serialVersionUID = -870190454845379171L;